using size_t instead of uint
git-svn-id: https://gearbox.svn.sf.net/svnroot/gearbox/gearbox/trunk@159 4b7f92db-2845-0410-95f1-b1f669fef64b

+++ src/gbxsickacfr/driver.cpp
@@ -379,7 +379,7 @@ Driver::guessLaserBaudRate()
     if ( config_.baudRate != 38400 ) baudRates.push_back( 38400 );
     if ( config_.baudRate != 500000 ) baudRates.push_back( 500000 );
     
-    for ( uint baudRateI=0; baudRateI < baudRates.size(); baudRateI++ )
+    for ( size_t baudRateI=0; baudRateI < baudRates.size(); baudRateI++ )
     {
         stringstream ss;
         ss << "Driver: Trying to connect at " << baudRates[baudRateI] << " baud.";
